encapsulate migration context
